Tests/LibWeb/Text/expected/WebAnimations/misc/keyframe-effect-cloning.txt

anim count: 1
keyframe: {"offset":null,"computedOffset":1,"easing":"linear","composite":"auto","marginLeft":"10px"}
